patch: ingress testing and rafactoring
* move a few files around and refactored code
* configured ingress testing and fixed some bugs in the PHP code that occured during migration from Helm charts
* updated the all stage - untested as of the moment
* started work on the monitoring tests
