+++ pkg/cmd/delete/delete_istio.go
@@ -1,6 +1,7 @@
 package delete
 
 import (
+	"fmt"
 	"github.com/spf13/cobra"
 	"k8s.io/kubectl/pkg/cmd/util"
 )
@@ -12,7 +13,7 @@ func NewDeleteIstioCmd(util.Factory) *cobra.Command {
 		Short: "Deletes the Istio testing resources",
 		Long:  `Deletes the Istio application testing suite.`,
 		Run: func(cmd *cobra.Command, args []string) {
-
+			fmt.Println("Istio will be supported soon™")
 		},
 	}
