+++ pkg/cmd/delete/delete_monitoring.go
@@ -1,20 +1,19 @@
 package delete
 
 import (
-	"github.com/drew-viles/k8s-e2e-tester/pkg/workloads/sql"
-	"github.com/drew-viles/k8s-e2e-tester/pkg/workloads/web"
+	"fmt"
 	"github.com/spf13/cobra"
 	"k8s.io/kubectl/pkg/cmd/util"
 	"log"
 )
 
-func NewDeleteCoreCmd(f util.Factory) *cobra.Command {
+func NewDeleteMonitoringCmd(f util.Factory) *cobra.Command {
 	o := &deleteOptions{}
 
 	cmd := &cobra.Command{
-		Use:   "core",
-		Short: "Delete a core workload",
-		Long:  "Deletes all elements of the core testing suite.",
+		Use:   "monitoring",
+		Short: "Deletes the Monitoring testing resources",
+		Long:  `Deletes the Monitoring application testing suite.`,
 		Run: func(cmd *cobra.Command, args []string) {
 			var err error
 
@@ -29,9 +28,7 @@ func NewDeleteCoreCmd(f util.Factory) *cobra.Command {
 				namespace = cmd.Flag("namespace").Value.String()
 			}
 
-			web.DeleteNginxWorkloadItems(o.client, namespace)
-			sql.DeleteSQLWorkloadItems(o.client, namespace)
-
+			fmt.Println(namespace)
 		},
 	}
