+++ pkg/cmd/validate/validate.go
@@ -1,14 +1,11 @@
 package validate
 
 import (
-	"github.com/docker/distribution/context"
 	promclient "github.com/prometheus-operator/prometheus-operator/pkg/client/versioned"
 	promscheme "github.com/prometheus-operator/prometheus-operator/pkg/client/versioned/scheme"
 	"github.com/spf13/cobra"
 	istioclient "istio.io/client-go/pkg/clientset/versioned"
 	istioscheme "istio.io/client-go/pkg/clientset/versioned/scheme"
-	apiv1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/kubernetes/scheme"
 	"k8s.io/kubectl/pkg/cmd/util"
@@ -16,7 +13,9 @@ import (
 )
 
 var (
-	storageClassFlag string
+	storageClassFlag  string
+	requestCPUFlag    string
+	requestMemoryFlag string
 )
 
 type validateOptions struct {
@@ -42,31 +41,25 @@ the required test suite will run against the resources to ensure everything is w
 		newValidateMonitoringCmd(f),
 	}
 
-	cmd.PersistentFlags().StringVar(&storageClassFlag, "storage-class", "longhorn", "Used to define the name of the storage class to use for Persistent Volumes.")
-
 	cmd.AddCommand(commands...)
 
 	return cmd
 }
 
-func createNamespaceIfNotExists(client *kubernetes.Clientset, namespace string) {
-	if namespace == "default" {
-		return
-	}
-	_, err := client.CoreV1().Namespaces().Get(context.Background(), namespace, metav1.GetOptions{})
-	if err == nil {
-		return
-	}
+// addCoreFlags adds the flags required for the core workload. It can't be persistent as some just won't be required for all tests.
+func addCoreFlags(cmd *cobra.Command) {
+	cmd.Flags().StringVar(&storageClassFlag, "storage-class", "", "The name of the storage class to use for Persistent Volumes")
+	cmd.Flags().StringVar(&requestCPUFlag, "cpu", "1", "The request CPU value to ensure scaling happens")
+	cmd.Flags().StringVar(&requestMemoryFlag, "memory", "1Gi", "The request memory value to ensure scaling happens")
+}
 
-	log.Println(err)
-	log.Printf("creating namespace: %s\n", namespace)
-
-	_, err = client.CoreV1().Namespaces().Create(context.Background(), &apiv1.Namespace{
-		ObjectMeta: metav1.ObjectMeta{
-			Name: namespace,
-		},
-	}, metav1.CreateOptions{})
-
+// addIngressFlags adds the flags required for the ingress workload. It can't be persistent as some just won't be required for all tests.
+func addIngressFlags(cmd *cobra.Command) {
+	cmd.Flags().StringVar(&ingressClassFlag, "ingress-class", "nginx", "The IngressClass name")
+	cmd.Flags().BoolVar(&enableTLSFlag, "enable-tls", false, "Whether to enable TLS on the Ingress endpoint. You must have cert-manager enabled and configured for this test to succeed")
+	cmd.Flags().StringVar(&annotationsFlag, "annotations", "", "Any additional annotations to add to the ingress resource in the format 'a=1,b=2'")
+	cmd.Flags().StringVar(&hostFlag, "host", "", "The fqdn for the ingress resource")
+	err := cmd.MarkFlagRequired("host")
 	if err != nil {
 		log.Fatalln(err)
 	}
