+++ pkg/workloads/coreworkloads/.gitignore
@@ -1,4 +1,5 @@
 #These will be added to the project and removed from this ignore file as they are validated and tested.
+daemonset.go
 gateway.go
 destinationrule.go
 job.go
